+++ src/ploomber/cli/__init__.py
@@ -1,7 +1,23 @@
-from ploomber.cli import (build, plot, task, report, interact, status,
-                          examples, install)
+from ploomber.cli import (
+    build,
+    plot,
+    task,
+    report,
+    interact,
+    status,
+    examples,
+    install,
+    nb,
+)
 
 __all__ = [
-    'task', 'plot', 'build', 'report', 'interact', 'status', 'examples',
-    'install'
+    'task',
+    'plot',
+    'build',
+    'report',
+    'interact',
+    'status',
+    'examples',
+    'install',
+    'nb',
 ]