+++ src/ploomber/tasks/taskfactory.py
@@ -1,7 +1,6 @@
 class TaskFactory:
     """Utility class for reducing boilerplate code
     """
-
     def __init__(self, task_class, product_class, dag):
         self.task_class = task_class
         self.product_class = product_class