+++ tests/tasks/test_sqlupload.py
@@ -7,8 +7,10 @@ import pandas as pd
 import pytest
 
 from ploomber import DAG
-from ploomber.products import PostgresRelation, File, GenericSQLRelation
+from ploomber.products import (PostgresRelation, File, GenericSQLRelation,
+                               SQLiteRelation)
 from ploomber.tasks import SQLUpload, PythonCallable
+from ploomber.clients import SQLAlchemyClient
 
 
 def make_data(product):
@@ -19,8 +21,7 @@ def make_data(product):
 @pytest.mark.parametrize('serializer, task_arg',
                          [('to_parquet', 'data.parquet'),
                           ('to_parquet', Path('data.parquet')),
-                          ('to_csv', 'data.csv')
-                          ])
+                          ('to_csv', 'data.csv')])
 def test_can_upload_a_file(serializer, task_arg, tmp_directory,
                            pg_client_and_schema):
     pg_client, schema = pg_client_and_schema
@@ -34,9 +35,8 @@ def test_can_upload_a_file(serializer, task_arg, tmp_directory,
     dag.clients[PostgresRelation] = pg_client
 
     SQLUpload(task_arg,
-              product=PostgresRelation((schema,
-                                        'test_can_upload_a_file',
-                                        'table')),
+              product=PostgresRelation(
+                  (schema, 'test_can_upload_a_file', 'table')),
               dag=dag,
               name='upload',
               to_sql_kwargs={'if_exists': 'replace'})
@@ -47,8 +47,7 @@ def test_can_upload_a_file(serializer, task_arg, tmp_directory,
 @pytest.mark.parametrize('serializer, task_arg',
                          [('to_parquet', 'data.parquet'),
                           ('to_parquet', Path('data.parquet')),
-                          ('to_csv', 'data.csv')
-                          ])
+                          ('to_csv', 'data.csv')])
 def test_upload_a_file_with_generic_relation(serializer, task_arg,
                                              sqlite_client_and_tmp_dir,
                                              pg_client_and_schema):
@@ -64,9 +63,8 @@ def test_upload_a_file_with_generic_relation(serializer, task_arg,
     dag.clients[GenericSQLRelation] = client
 
     SQLUpload(task_arg,
-              product=GenericSQLRelation((schema,
-                                          'test_can_upload_a_file',
-                                          'table')),
+              product=GenericSQLRelation(
+                  (schema, 'test_can_upload_a_file', 'table')),
               dag=dag,
               name='upload',
               to_sql_kwargs={'if_exists': 'replace'})
@@ -86,16 +84,20 @@ def test_append_rows(tmp_directory, pg_client_and_schema):
     dag.clients[PostgresRelation] = pg_client
 
     # create table
-    df.to_sql('test_append', pg_client.engine,
-              schema=schema, if_exists='replace', index=False)
+    df.to_sql('test_append',
+              pg_client.engine,
+              schema=schema,
+              if_exists='replace',
+              index=False)
 
     SQLUpload('data.csv',
-              product=PostgresRelation((schema,
-                                        'test_append',
-                                        'table')),
+              product=PostgresRelation((schema, 'test_append', 'table')),
               dag=dag,
               name='upload',
-              to_sql_kwargs={'if_exists': 'append', 'index': False})
+              to_sql_kwargs={
+                  'if_exists': 'append',
+                  'index': False
+              })
 
     dag.build()
 
@@ -122,9 +124,7 @@ def test_can_upload_file_from_upstream_dependency(tmp_directory,
 
     name = 'test_can_upload_file_from_upstream_dependency'
     pg = SQLUpload('{{upstream["make"]}}',
-                   product=PostgresRelation((schema,
-                                             name,
-                                             'table')),
+                   product=PostgresRelation((schema, name, 'table')),
                    dag=dag,
                    name='upload',
                    to_sql_kwargs={'if_exists': 'replace'})
@@ -132,3 +132,31 @@ def test_can_upload_file_from_upstream_dependency(tmp_directory,
     make >> pg
 
     dag.build()
+
+
+def test_custom_io_handler(tmp_directory):
+    dag = DAG()
+    client = SQLAlchemyClient('sqlite:///database.db')
+    dag.clients[SQLUpload] = client
+    dag.clients[SQLiteRelation] = client
+
+    df = pd.DataFrame({'a': [1, 2, 3], 'b': [1, 2, 3]})
+    df.to_csv('some-file.tsv', sep='\t', index=False)
+
+    def my_reading_fn(path):
+        return pd.read_csv(path, sep='\t')
+
+    SQLUpload('some-file.tsv',
+              SQLiteRelation(('my-table', 'table')),
+              dag=dag,
+              name='task',
+              io_handler=my_reading_fn,
+              to_sql_kwargs=dict(index=False))
+
+    dag.build()
+
+    other = pd.read_sql('SELECT * FROM "my-table"', con=client)
+
+    client.close()
+
+    assert other.equals(df)